requirements.txt
appdirs==1.4.4
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
daraja-mpesa==1.2.1
distlib==0.3.1
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
filelock==3.0.12
future==0.18.2
idna==3.3
importlib-metadata==4.11.2
Markdown==3.3.6
pygame==1.9.6
python-decouple==3.6
pytz==2021.3
requests==2.27.1
six==1.15.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.4.4
virtualenvwrapper-win==1.2.6
zipp==3.7.0
